pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ladderbot"
version = "0.1.0"
description = "Sports betting parlay optimizer with ladder tracking"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "nba_api>=1.4.0",
    "schedule>=1.2.0",
    "sse-starlette>=1.6.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.21.0"]

[tool.setuptools.packages.find]
include = ["ladderbot*"]

[project.scripts]
ladderbot = "ladderbot.run:main"
